+++ src/main/java/fiji/plugin/trackmate/util/cli/CondaCLIConfigurator.java
@@ -124,7 +124,7 @@ public abstract class CondaCLIConfigurator extends CLIConfigurator
 		condaEnv.set( 0 );
 
 		// Add the translator to make a proper cmd line calling conda first.
-		setTranslator( condaEnv, s -> {
+		setCommandTranslator( condaEnv, s -> {
 			final List< String > cmd = new ArrayList<>();
 			final String condaPath = CLIUtils.getCondaPath();
 			// Conda and executable stuff.
